+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.8 branch
 ### 2.8.6
+* fixed bootstrap, `mkdir` errors are no longer suppressed;
 * extensive improvement of function descriptions and tags. The level of `phpstan`
     has been raised.